Refactor Hammerspoon config into modules
